changed long to int for no_elements

+++ lib/THCLNN/SpatialUpSamplingNearest.cl
@@ -35,7 +35,7 @@
 
 }
 
-kernel void upscale(global float *input_data, int input_offset, global float *output_data, int output_offset, long no_elements,
+kernel void upscale(global float *input_data, int input_offset, global float *output_data, int output_offset, int no_elements,
                         int scale_factor, int d1, int d2, int d3)
 {
   global float *input = input_data + input_offset;
@@ -51,7 +51,7 @@ kernel void upscale(global float *input_data, int input_offset, global float *ou
 /*
  * Description:
  */
-kernel void downscale(global float *gradInput_data_data, int gradInput_data_offset, global float *gradOutput_data_data, int gradOutput_data_offset, long no_elements,
+kernel void downscale(global float *gradInput_data_data, int gradInput_data_offset, global float *gradOutput_data_data, int gradOutput_data_offset, int no_elements,
                               int scale_factor, int d1, int d2, int d3)
 {
   global float *gradInput_data = gradInput_data_data + gradInput_data_offset;
